requirements.txt
langchain==0.1.20
langchain-openai==0.1.8
langchain-google-genai==0.0.15
langchain-community==0.0.38
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.0
google-generativeai==0.3.0
